Add Formatter.java class.
